Formatando os dados dos demais endpoints

+++ CalcInvest/CalcInvest.Application/JurosCompostoDTO.cs
@@ -3,10 +3,10 @@
     public class JurosCompostosDTO
     {
 
-        public double CapitalInicial { get; set; }
-        public double TaxaDeJuros { get; set; }
+        public decimal CapitalInicial { get; set; }
+        public decimal TaxaDeJuros { get; set; }
         public string AnualouMensal { get; set; }
-        public double ValorMensal { get; set; } 
+        public decimal ValorMensal { get; set; } 
         public int TempoMeses { get; set; }
