+++ CalcInvest/CalcInvest.Application/JurosSimplesDTO.cs
@@ -3,9 +3,9 @@
     public class JurosSimplesDTO
     {
 
-        public double CapitalInicial { get; set; }
+        public decimal CapitalInicial { get; set; }
 
-        public double TaxaDeJuros { get; set; }
+        public decimal TaxaDeJuros { get; set; }
 
         public string AnualouMensal { get; set; }
