+++ CalcInvest/CalcInvest.Core/CalcJurosSimples.cs
@@ -2,27 +2,27 @@
 {
     public class CalcJurosSimples
     {
-        public double CapitalInicial { get; set; }
+        public decimal CapitalInicial { get; set; }
 
         public TaxaDeJuros TaxaDeJuros { get; set; }
 
         public int TempoMeses { get; set; }
 
-        public double ValorFinal { get; set; }
+        public decimal ValorFinal { get; set; }
 
-        public CalcJurosSimples(double valorInicial, TaxaDeJuros taxaDeJuros,  int meses)
+        public CalcJurosSimples(decimal valorInicial, TaxaDeJuros taxaDeJuros,  int meses)
         {
             CapitalInicial = valorInicial;
             TaxaDeJuros = taxaDeJuros;
             TempoMeses = meses;
         }
 
-        public double CalcularJurosSimples()
+        public decimal CalcularJurosSimples()
         {
 
             if(TaxaDeJuros.Tipo == "Mensal" || TaxaDeJuros.Tipo == "mensal")
             {
-                double taxaDeJuros = TaxaDeJuros.Porcentagem/100;
+                decimal taxaDeJuros = TaxaDeJuros.Porcentagem/100;
 
 
                 ValorFinal = CapitalInicial * taxaDeJuros * TempoMeses;
@@ -31,9 +31,9 @@
                 return ValorFinal;
             }
 
-            double taxaDeJurosMensal = TaxaDeJuros.Porcentagem / 12;
+            decimal taxaDeJurosMensal = TaxaDeJuros.Porcentagem / 12;
 
-            double taxaDeJursoFinal = taxaDeJurosMensal/100;
+            decimal taxaDeJursoFinal = taxaDeJurosMensal/100;
 
             ValorFinal = CapitalInicial * taxaDeJursoFinal * TempoMeses;
